Dropping in a replacement layout for activity_main.
When the fragment class is named in the layout, replace is the only possible transaction. When the activity is showing both fragments, there are no transactions. An orientation change might change the width of a device and switching fragments could happen at that time. Using a view group container, as seen in the master-detail flow template, the chessstar app can use add and show fragment transactions to be more efficient